+++ lib/freydis/cryptsetup.rb
@@ -3,27 +3,74 @@
 module Freydis
   class Cryptsetup
     def initialize(data)
-      raise StandardError, "Bad uuid #{uuid}" unless File.exists? "/dev/disk/by-uuid/#{uuid}"
-      @uuid = uuid
-      @name = "freydis-enc"
+      @data = data
+      @dev_ids = [
+        "/dev/disk/by-id/" + @data.options[:disk_id] ||= " ",
+        "/dev/disk/by-uuid/" + @data.options[:disk_uuid] ||= " ",
+        "/dev/disk/by-partuuid/" + @data.options[:disk_partuuid] ||= " ",
+        "/dev/" + @data.options[:disk]
+      ]
+      @mapper_name = "freydis-enc"
+      @mountpoint ="/mnt/freydis"
     end
 
     def encrypt
-      exec "cryptsetup -v --type luks2 --verify-passphrase luksFormat /dev/disk/by-uuid/#{@uuid}"
-      raise StandardError, "encrypt #{@uuid}" unless $?.exitstatus == 0
+      @dev_ids.each { |f|
+        if File.exists? f
+          exec "cryptsetup -v --type luks2 --verify-passphrase luksFormat #{f}"
+          break if $?.success?
+        end
+      }
     end
 
     def open
-      `sudo cryptsetup -v open "#{@uuid}" #{@name}`
+      @dev_ids.each { |f|
+        if File.exists? f
+          exec "cryptsetup -v open #{f} #{@mapper_name}"
+          break if $?.success?
+        end
+      }
     end
 
     def close
-      `sudo cryptsetup -v close #{@name}`
+      umount
+      exec "cryptsetup -v close #{@mapper_name}" if File.exists? "/dev/mapper/#{@mapper_name}"
+    end
+
+    def format
+      exec "mkfs.ext4 /dev/mapper/#{@mapper_name}"
+    end
+
+    def mount
+      create_mountpoint
+      puts "Mounting disk at #{@mountpoint}"
+      exec "mount -t ext4 /dev/mapper/#{@mapper_name} #{@mountpoint}"
     end
 
     private
+
+    def create_mountpoint
+      if Process.uid === 0
+        Dir.mkdir @mountpoint unless Dir.exist? @mountpoint
+      else
+        exec "mkdir -p #{@mountpoint}" unless Dir.exist? @mountpoint
+      end
+    end
+
+    def umount
+      dir_length = Dir.glob(@mountpoint).length
+      if dir_length >= 1 # should contain lost+found if mount
+        exec "umount #{@mountpoint}"
+      end
+    end
+
     def exec(command)
-      system("sudo #{command}")
+      sudo = Process.uid != 0 ? 'sudo' : ''
+      if system("#{sudo} #{command}")
+        puts "#{command} done."
+      else
+        raise StandardError, "[-] #{command}"
+      end
     end
   end
 end