+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+* Can (u)mount the encrypted device (default /mnt/freydis)
+* Checking all ID (partuuid,uuid,id) from a given device
 * YAML config file in ~/.config/freydis/freydis.yaml
 
 ## 0.0.1, release 05/04/21